+++ src/event-listeners/ChangeEventListener.test.ts
@@ -37,7 +37,7 @@ describe('ChangeEventListener', () => {
     fireEvent.change(search)
 
     await waitFor(() => {
-      expect(createTargetedUserActionSpy).toHaveBeenCalledWith(new Event('change'), 'change', /.*/, undefined)
+      expect(createTargetedUserActionSpy).toHaveBeenCalledWith(new Event('change'), 'change', { excludeRegex: /.*/ })
     })
   })
 
@@ -56,7 +56,9 @@ describe('ChangeEventListener', () => {
     fireEvent.change(search)
 
     await waitFor(() => {
-      expect(createTargetedUserActionSpy).toHaveBeenCalledWith(new Event('change'), 'change', undefined, 'data-testid')
+      expect(createTargetedUserActionSpy).toHaveBeenCalledWith(new Event('change'), 'change', {
+        customSelector: 'data-testid',
+      })
     })
   })