+++ src/event-listeners/KeyUpEventListener.ts
@@ -10,12 +10,7 @@ class KeyUpEventListener extends TargetedEventListener {
   }
 
   listener(event: KeyboardEvent) {
-    const userAction = createTargetedUserAction(
-      event,
-      this.userActionType,
-      this.options.excludeRegex,
-      this.options.customSelector,
-    )
+    const userAction = createTargetedUserAction(event, this.userActionType, this.options)
     if (userAction === null) return
     if (isKeyAllowedByKeyListeners(event.code)) {
       return this.userActionHandler.handle(userAction)